fix fatal errors on PHP 5.6
